+++ src/RadixDlt.NetworkGateway.Abstractions/StandardMetadata/StandardMetadataConstants.cs
@@ -1,0 +1,12 @@
+namespace RadixDlt.NetworkGateway.Abstractions.StandardMetadata;
+
+public static class StandardMetadataConstants
+{
+    public const string DappAccountType = "account_type";
+    public const string DappDefinition = "dapp_definition";
+    public const string DappDefinitions = "dapp_definitions";
+    public const string DappClaimedWebsites = "claimed_websites";
+    public const string DappClaimedEntities = "claimed_entities";
+
+    public const string DappAccountTypeDappDefinition = "dapp definition";
+}