+++ src/RadixDlt.NetworkGateway.PostgresIntegration/CustomTypes.cs
@@ -65,7 +65,7 @@
 using Dapper;
 using Npgsql;
 using RadixDlt.NetworkGateway.Abstractions.Model;
-using RadixDlt.NetworkGateway.Abstractions.TwoWayLinks;
+using RadixDlt.NetworkGateway.Abstractions.StandardMetadata;
 using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 using RadixDlt.NetworkGateway.PostgresIntegration.ValueConverters;
 using NonFungibleIdType = RadixDlt.NetworkGateway.Abstractions.Model.NonFungibleIdType;
@@ -121,7 +121,7 @@ internal static class CustomTypes
         NpgsqlConnection.GlobalTypeMapper.MapEnum<SborTypeKind>();
         NpgsqlConnection.GlobalTypeMapper.MapEnum<StateType>();
         NpgsqlConnection.GlobalTypeMapper.MapEnum<AuthorizedDepositorBadgeType>();
-        NpgsqlConnection.GlobalTypeMapper.MapEnum<TwoWayLinkType>();
+        NpgsqlConnection.GlobalTypeMapper.MapEnum<StandardMetadataKey>();
 #pragma warning restore CS0618
 
         _configured = true;