+++ src/RadixDlt.NetworkGateway.PostgresIntegration/ServiceCollectionExtensions.cs
@@ -69,7 +69,7 @@ using Microsoft.Extensions.DependencyInjection.Extensions;
 using Microsoft.Extensions.Logging;
 using Npgsql;
 using RadixDlt.NetworkGateway.Abstractions.Model;
-using RadixDlt.NetworkGateway.Abstractions.TwoWayLinks;
+using RadixDlt.NetworkGateway.Abstractions.StandardMetadata;
 using RadixDlt.NetworkGateway.PostgresIntegration.Models;
 
 namespace RadixDlt.NetworkGateway.PostgresIntegration;
@@ -121,7 +121,7 @@ public static class ServiceCollectionExtensions
                 dataSourceBuilder.MapEnum<SborTypeKind>();
                 dataSourceBuilder.MapEnum<StateType>();
                 dataSourceBuilder.MapEnum<AuthorizedDepositorBadgeType>();
-                dataSourceBuilder.MapEnum<TwoWayLinkType>();
+                dataSourceBuilder.MapEnum<StandardMetadataKey>();
 
                 return new NpgsqlDataSourceHolder<T>(dataSourceBuilder.Build());
             },